pyproject.toml
[project]
name = "kalshi_decision_engine"
version = "0.1.0"
description = "SAFE, compliant Kalshi algorithmic trading bot with risk-managed decisions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Codex"}]

dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.26",
  "websockets>=12.0",
  "cryptography>=42.0",
  "rich>=13.7",
  "typer>=0.12",
  "PyYAML>=6.0",
  "kalshi-python>=0.1.1",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
async = ["kalshi-python-async>=3.0.0"]
sync = ["kalshi-python-sync>=3.0.0"]

test = ["pytest>=8.0"]

[project.scripts]
kalshi-bot = "kalshi_bot.cli:app"

[tool.setuptools.packages.find]
include = ["kalshi_bot*"]

[tool.pytest.ini_options]
addopts = "-q"
